+++ storage/migrations/2024_01_01_000017_add_printer_to_details.sql
@@ -3,7 +3,7 @@
 -- =====================================================
 
 ALTER TABLE details
-    ADD COLUMN printer_id INT UNSIGNED NULL AFTER material_item_id,
+    ADD COLUMN printer_id BIGINT UNSIGNED NULL AFTER material_item_id,
     ADD INDEX idx_printer_id (printer_id),
     ADD CONSTRAINT fk_details_printer_id
         FOREIGN KEY (printer_id) REFERENCES printers(id) ON DELETE SET NULL;
